Removed Alias feature.
Changed converters to accept null values.  Needed a clean way to create new values for non-existant fields in Records.  Concat/Constant operations were not being executed if the starting value was null.  Converters not return null if null is passed in, except for the Concat/Constant converters (and a few others) that allow null input.

git-svn-id: svn+ssh://dev.real-comp.com/opt/usr/share/rcsvn/realcomp-data/branches/data-type-operations/realcomp-data@3151 cb9be8fc-bb24-4f4f-8dc9-9fc556ea96d8