removed org.reflections dependency and replaced it with io.github.luckhutch fast-classpath-scanner.  org.reflections was stuck on an old version of Guava and it was causing issues in a Hadoop M/R environment.
Removed the 'match' feature of the field list and un-deprecated the 'classifier' property.  This was a temporary experiment to allow FieldList classification on a regex pattern that didn't have to match the whole record.  I regret that the classifier uses the Matcher.match() method instead of Matcher.find() which would have been more powerful.  Changing this would break backward compatibility (e.g., mvr tx schemas) so not doing this at this time.